+++ include/visionaray/detail/spd/d65.h
@@ -28,7 +28,7 @@ public:
             return 0;
         }
 
-        auto i = (int)floorf(lambda) / 10;
+        auto i = (int)floor(lambda) / 10;
         auto s = (lambda - i * 10.0f) / 10.0f;
 
         i -= 30;
